Added ScanRecordParser class and removed UUIDUtil
