+++ mockrxandroidble/src/test/groovy/com/polidea/rxandroidble2/mockrxandroidble/RxBleClientMockTest.groovy
@@ -10,6 +10,7 @@ import hkhc.electricspock.ElectricSpecification
 import io.reactivex.Observable
 import io.reactivex.subjects.PublishSubject
 import org.robolectric.annotation.Config
+import com.polidea.rxandroidble2.BuildConfig
 
 @Config(manifest = Config.NONE, constants = BuildConfig, sdk = Build.VERSION_CODES.LOLLIPOP)
 public class RxBleClientMockTest extends ElectricSpecification {
