+++ rxandroidble/src/main/java/com/polidea/rxandroidble2/RxBleClientImpl.java
@@ -20,6 +20,7 @@ import com.polidea.rxandroidble2.internal.util.CheckerLocationPermission;
 import com.polidea.rxandroidble2.internal.util.ClientStateObservable;
 import com.polidea.rxandroidble2.internal.util.LocationServicesStatus;
 import com.polidea.rxandroidble2.internal.util.RxBleAdapterWrapper;
+import com.polidea.rxandroidble2.internal.util.ScanRecordParser;
 import com.polidea.rxandroidble2.scan.BackgroundScanner;
 import com.polidea.rxandroidble2.scan.ScanFilter;
 import com.polidea.rxandroidble2.scan.ScanResult;
@@ -52,6 +53,7 @@ class RxBleClientImpl extends RxBleClient {
     @Deprecated
     public static final String TAG = "RxBleClient";
     final ClientOperationQueue operationQueue;
+    private final ScanRecordParser scanRecordParser;
     private final RxBleDeviceProvider rxBleDeviceProvider;
     final ScanSetupBuilder scanSetupBuilder;
     final ScanPreconditionsVerifier scanPreconditionVerifier;
@@ -67,10 +69,10 @@ class RxBleClientImpl extends RxBleClient {
     private final CheckerLocationPermission checkerLocationPermission;
 
     @Inject
-    @SuppressWarnings("deprecation")
     RxBleClientImpl(RxBleAdapterWrapper rxBleAdapterWrapper,
                     ClientOperationQueue operationQueue,
                     Observable<BleAdapterState> adapterStateObservable,
+                    ScanRecordParser scanRecordParser,
                     LocationServicesStatus locationServicesStatus,
                     Lazy<ClientStateObservable> lazyClientStateObservable,
                     RxBleDeviceProvider rxBleDeviceProvider,
@@ -84,6 +86,7 @@ class RxBleClientImpl extends RxBleClient {
         this.operationQueue = operationQueue;
         this.rxBleAdapterWrapper = rxBleAdapterWrapper;
         this.rxBleAdapterStateObservable = adapterStateObservable;
+        this.scanRecordParser = scanRecordParser;
         this.locationServicesStatus = locationServicesStatus;
         this.lazyClientStateObservable = lazyClientStateObservable;
         this.rxBleDeviceProvider = rxBleDeviceProvider;
@@ -149,7 +152,7 @@ class RxBleClientImpl extends RxBleClient {
     }
 
     @Override
-    @SuppressWarnings("deprecation")
+    @Deprecated
     public Observable<RxBleScanResult> scanBleDevices(@Nullable final UUID... filterServiceUUIDs) {
         return Observable.defer(new Callable<ObservableSource<? extends RxBleScanResult>>() {
             @Override
@@ -211,7 +214,7 @@ class RxBleClientImpl extends RxBleClient {
     private Observable<RxBleScanResult> createScanOperationApi18(@Nullable final UUID[] filterServiceUUIDs) {
         final Set<UUID> filteredUUIDs = toDistinctSet(filterServiceUUIDs);
         final LegacyScanOperation
-                scanOperation = new LegacyScanOperation(filterServiceUUIDs, rxBleAdapterWrapper);
+                scanOperation = new LegacyScanOperation(filterServiceUUIDs, rxBleAdapterWrapper, scanRecordParser);
         return operationQueue.queue(scanOperation)
                 .doFinally(new Action() {
                     @Override
