+++ rxandroidble/src/test/groovy/com/polidea/rxandroidble2/RxBleClientTest.groovy
@@ -2,6 +2,7 @@ package com.polidea.rxandroidble2
 
 import android.bluetooth.BluetoothDevice
 import android.content.Context
+import android.os.Build
 import com.polidea.rxandroidble2.exceptions.BleScanException
 
 import com.polidea.rxandroidble2.internal.RxBleDeviceProvider
@@ -10,26 +11,30 @@ import com.polidea.rxandroidble2.internal.scan.*
 import com.polidea.rxandroidble2.internal.serialization.ClientOperationQueue
 import com.polidea.rxandroidble2.internal.util.CheckerLocationPermission
 import com.polidea.rxandroidble2.internal.util.ClientStateObservable
+import com.polidea.rxandroidble2.internal.util.ScanRecordParser
 import com.polidea.rxandroidble2.scan.BackgroundScanner
 import com.polidea.rxandroidble2.scan.ScanSettings
+import hkhc.electricspock.ElectricSpecification
 import io.reactivex.Observable
 import io.reactivex.ObservableSource
 import io.reactivex.ObservableTransformer
 import io.reactivex.annotations.NonNull
 import io.reactivex.observers.TestObserver
 import io.reactivex.schedulers.TestScheduler
-import spock.lang.Specification
+import org.robolectric.annotation.Config
 import spock.lang.Unroll
 
 import static com.polidea.rxandroidble2.exceptions.BleScanException.*
 
 @SuppressWarnings("GrDeprecatedAPIUsage")
-class RxBleClientTest extends Specification {
+@Config(manifest = Config.NONE, constants = BuildConfig, sdk = Build.VERSION_CODES.LOLLIPOP)
+class RxBleClientTest extends ElectricSpecification {
 
     BackgroundScanner backgroundScanner = Mock(BackgroundScanner)
     DummyOperationQueue dummyQueue = new DummyOperationQueue()
     RxBleClient objectUnderTest
     Context contextMock = Mock Context
+    ScanRecordParser scanRecordParserSpy = Spy ScanRecordParser
     MockRxBleAdapterWrapper bleAdapterWrapperSpy = Spy MockRxBleAdapterWrapper
     MockRxBleAdapterStateObservable adapterStateObservable = Spy MockRxBleAdapterStateObservable
     MockLocationServicesStatus locationServicesStatusMock = Spy MockLocationServicesStatus
@@ -75,6 +80,7 @@ class RxBleClientTest extends Specification {
                 bleAdapterWrapperSpy,
                 queue,
                 adapterStateObservable.asObservable(),
+                scanRecordParserSpy,
                 locationServicesStatusMock,
                 mockLazyClientStateObservable,
                 mockDeviceProvider,
@@ -410,7 +416,7 @@ class RxBleClientTest extends Specification {
     def "should emit devices only if matching filter (#description)"() {
         given:
         addressList.each { bluetoothDeviceDiscovered deviceMac: it, rssi: 0, scanRecord: [] as byte[] }
-        uuidParserSpy.extractUUIDs(_) >>> publicServices
+        scanRecordParserSpy.extractUUIDs(_) >>> publicServices
 
         when:
         TestObserver<RxBleScanResult> testSubscriber = objectUnderTest.scanBleDevices(filter as UUID[]).test()
@@ -444,7 +450,7 @@ class RxBleClientTest extends Specification {
         def secondUUID = UUID.randomUUID()
         def thirdUUID = UUID.randomUUID()
         bluetoothDeviceDiscovered deviceMac: "AA:AA:AA:AA:AA:AA", rssi: 0, scanRecord: [] as byte[]
-        uuidParserSpy.extractUUIDs(_) >> [filter, secondUUID, thirdUUID]
+        scanRecordParserSpy.extractUUIDs(_) >> [filter, secondUUID, thirdUUID]
 
         when:
         def testSubscriber = objectUnderTest.scanBleDevices([filter] as UUID[]).test()
